requirements.txt
# Core Dependencies
earthengine-api>=0.1.370
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
numpy>=1.24.0
scikit-image>=0.21.0
scipy>=1.10.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Database (optional)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
geoalchemy2>=0.14.0

# Data Processing
pandas>=2.0.0
xarray>=2023.1.0
dask[complete]>=2023.1.0

# Visualization & Dashboard
folium>=0.14.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# Notifications  
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0